requirements.txt
openai
python-dotenv
bs4
requests
